oculus-5.2.0-v55-automation @Commit ad6bc5cb96acfcad05e5b5d45df6c7da262a24cc
OVRPlugin 1.87.0
Platform SDK 1.76.0
Audio SDK 1.41.0

+++ Engine/Plugins/AI/MassCrowd/Source/MassCrowd/Private/MassCrowdReplicator.cpp
@@ -1,14 +1,8 @@
 // Copyright Epic Games, Inc. All Rights Reserved.
 
 #include "MassCrowdReplicator.h"
-#include "MassReplicationTypes.h"
-#include "MassClientBubbleHandler.h"
 #include "MassCrowdBubble.h"
-#include "MassCommonFragments.h"
-#include "MassLODSubsystem.h"
-#include "MassCrowdFragments.h"
-#include "MassReplicationPathHandlers.h"
-#include "MassReplicationTransformHandlers.h"
+#include "MassExecutionContext.h"
 
 //----------------------------------------------------------------------//
 //  UMassCrowdReplicator
@@ -35,7 +29,7 @@ void UMassCrowdReplicator::ProcessClientReplication(FMassExecutionContext& Conte
 		check(RepSharedFrag);
 	};
 
-	auto AddEntityCallback = [&RepSharedFrag, &PathHandler, &PositionYawHandler](FMassExecutionContext& Context,const int32 EntityIdx, FReplicatedCrowdAgent& InReplicatedAgent, const FMassClientHandle ClientHandle)->FMassReplicatedAgentHandle
+	auto AddEntityCallback = [&RepSharedFrag, &PathHandler, &PositionYawHandler](FMassExecutionContext& Context, const int32 EntityIdx, FReplicatedCrowdAgent& InReplicatedAgent, const FMassClientHandle ClientHandle)->FMassReplicatedAgentHandle
 	{
 		AMassCrowdClientBubbleInfo& CrowdBubbleInfo = RepSharedFrag->GetTypedClientBubbleInfoChecked<AMassCrowdClientBubbleInfo>(ClientHandle);
 
@@ -45,7 +39,7 @@ void UMassCrowdReplicator::ProcessClientReplication(FMassExecutionContext& Conte
 		return CrowdBubbleInfo.GetCrowdSerializer().Bubble.AddAgent(Context.GetEntity(EntityIdx), InReplicatedAgent);
 	};
 
-	auto ModifyEntityCallback = [&RepSharedFrag, &PathHandler](FMassExecutionContext& Context, const int32 EntityIdx, const EMassLOD::Type LOD, const float Time, const FMassReplicatedAgentHandle Handle, const FMassClientHandle ClientHandle)
+	auto ModifyEntityCallback = [&RepSharedFrag, &PathHandler](FMassExecutionContext& Context, const int32 EntityIdx, const EMassLOD::Type LOD, const double Time, const FMassReplicatedAgentHandle Handle, const FMassClientHandle ClientHandle)
 	{
 		AMassCrowdClientBubbleInfo& CrowdBubbleInfo = RepSharedFrag->GetTypedClientBubbleInfoChecked<AMassCrowdClientBubbleInfo>(ClientHandle);
