+++ Engine/Source/Editor/UnrealEd/Private/ThumbnailRendering/FontThumbnailRenderer.cpp
@@ -7,7 +7,9 @@
 #include "Engine/Font.h"
 #include "CanvasItem.h"
 #include "Engine/Texture2D.h"
+#include "CanvasItem.h"
 #include "CanvasTypes.h"
+#include "TextureResource.h"
 
 UFontThumbnailRenderer::UFontThumbnailRenderer(const FObjectInitializer& ObjectInitializer)
 	: Super(ObjectInitializer)
@@ -36,7 +38,7 @@ void UFontThumbnailRenderer::GetThumbnailSize(UObject* Object, float Zoom, uint3
 		case EFontCacheType::Runtime:
 			if(Font->CompositeFont.DefaultTypeface.Fonts.Num() > 0)
 			{
-				OutWidth = OutHeight = Zoom * 256.0f;
+				OutWidth = OutHeight = static_cast<uint32>(Zoom * 256.0f);
 			}
 			break;
